requirements.txt
numpy==1.23.4
# tensorflow
tensorflow-macos
tensorflow_hub
Pillow
matplotlib
pyyaml 
h5py
